requirements.txt
openai
python-dotenv
pydantic
fastapi
uvicorn
python-multipart
notion-client
